compile on 32 bit

+++ src/portaudio/pa.rs
@@ -5,6 +5,7 @@
 
 use std::{str, ptr, vec, mem, cast};
 use std::libc::{c_double, c_void, malloc};
+use std::libc::types::os::arch::c95::size_t;
 
 use types::*;
 use ffi;
@@ -329,7 +330,7 @@ impl<S> PaStream<S> {
         if !input_parameters.is_none() {
             self.c_input = Some(input_parameters.unwrap().unwrap());
             self.num_input_channels = input_parameters.unwrap().channel_count;
-            self.unsafe_buffer = unsafe { malloc(mem::size_of::<S>() as u64 * frames_per_buffer as u64 * input_parameters.unwrap().channel_count as u64) };
+            self.unsafe_buffer = unsafe { malloc(mem::size_of::<S>() as size_t * frames_per_buffer as size_t * input_parameters.unwrap().channel_count as size_t) };
         }
         if !output_parameters.is_none() {
             self.c_output = Some(output_parameters.unwrap().unwrap());
@@ -381,7 +382,7 @@ impl<S> PaStream<S> {
         if num_input_channels > 0 {
             self.c_input = None;
             self.num_input_channels = num_input_channels;
-            self.unsafe_buffer = unsafe { malloc(mem::size_of::<S>() as u64 * frames_per_buffer as u64 * num_input_channels as u64) };
+            self.unsafe_buffer = unsafe { malloc(mem::size_of::<S>() as size_t * frames_per_buffer as size_t * num_input_channels as size_t) };
         }
         unsafe {
            ffi::Pa_OpenDefaultStream(&self.c_pa_stream, num_input_channels, num_output_channels, sample_format, sample_rate as c_double, frames_per_buffer, None, ptr::null())
